fix: variables in template filenames
