+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
-  testMatch: ['**/src/__tests__/*.+(ts|tsx|js)', '**/tests/*.test.(ts|tsx|js)'],
+  testMatch: ['**/src/__tests__/**/*.+(ts|tsx|js)', '**/tests/**/*.test.(ts|tsx|js)'],
   testPathIgnorePatterns: ['/lib/'],
 };
