+++ src/template.ts
@@ -87,11 +87,15 @@ export interface CopyConfig {
 export async function copy(args: CopyConfig) {
   const templateFiles = await globby(slash(args.templateDir), { dot: true });
   for (const sourcePath of templateFiles) {
-    const relativePath = path.relative(args.templateDir, sourcePath);
-    const targetPath = slash(format(
-      path.resolve(args.packageDir, relativePath),
+    let relativePath = format(
+      path.relative(args.templateDir, sourcePath),     
       args.view
-    )).replace(/\/gitignore$/, '/.gitignore'); // https://github.com/uetchy/create-create-app/issues/38
+    );
+
+    let targetPath = slash(
+      path.resolve(args.packageDir, relativePath)
+    ).replace(/\/gitignore$/, '/.gitignore'); // https://github.com/uetchy/create-create-app/issues/38
+
     prepareDirectory(targetPath);
 
     let sourceData = fs.readFileSync(sourcePath);
