+++ tests/unit/test_templates/default/{{name}}.md
@@ -1,0 +1,1 @@
+{{name}}: testing variables on file names
